==> Sars_Cov_2/cod/39_mutate_aa.ipynb <==
import pandas as pd
import random
from collections import Counter
from tqdm import tqdm
import matplotlib.pyplot as plt
from matplotlib.backends.backend_pdf import PdfPages
import taxoniq
from Bio.SeqUtils import seq3
from Bio import Entrez

viruses_specs = pd.read_csv('../data/ms12all_all_virus.csv', index_col=0)
viruses_type = pd.read_csv('../data/taxid_virus_type.csv', index_col=0)
viruses_aa_freq = pd.read_csv('../data/aminoacid_freq_all_virus.csv')

viruses_specs.head()

viruses_type.head()

viruses_type[viruses_type['Type'] == '-']['taxid'].unique()

viruses_type['taxid'] = viruses_type['taxid'].apply(str)

viruses_aa_freq.head()

d = {'C': 'Cys', 'D': 'Asp', 'S': 'Ser', 'Q': 'Gln', 'K': 'Lys',
     'I': 'Ile', 'P': 'Pro', 'T': 'Thr', 'F': 'Phe', 'N': 'Asn', 
     'G': 'Gly', 'H': 'His', 'L': 'Leu', 'R': 'Arg', 'W': 'Trp', 
     'A': 'Ala', 'V':'Val', 'E': 'Glu', 'Y': 'Tyr', 'M': 'Met'}

viruses_aa_freq.columns = [d[col] if col in d else col for col in viruses_aa_freq.columns]

viruses_aa_freq.head()

viruses_aa_freq['taxid'].unique()

viruses_specs['taxid'].unique()

viruses_type['taxid'].unique()

viruses_aa_freq['taxid'] = viruses_aa_freq['taxid'].apply(str)

viruses_specs.loc[viruses_specs['taxid'] == '118655(1933309)','taxid'] = '118655'
viruses_type.loc[viruses_type['taxid'] == '1933309','taxid'] = '118655'

viruses_specs.loc[viruses_specs['taxid'] == '54290_1985356','taxid'] = '54290'
viruses_type.loc[viruses_type['taxid'] == '1985356','taxid'] = '54290'

codon_aa={'UUU':'Phe','UUC':'Phe','UUA':'Leu','UUG':'Leu','CUU':'Leu','CUC':'Leu','CUA':'Leu','CUG':'Leu','AUU':'Ile',
       'AUC':'Ile','AUA':'Ile','AUG':'Met','GUU':'Val','GUC':'Val','GUA':'Val','GUG':'Val','UCU':'Ser','UCC':'Ser',
       'UCA':'Ser','UCG':'Ser','CCU':'Pro','CCC':'Pro','CCA':'Pro','CCG':'Pro','ACU':'Thr','ACC':'Thr','ACA':'Thr',
       'ACG':'Thr','GCU':'Ala','GCC':'Ala','GCA':'Ala','GCG':'Ala','UAU':'Tyr','UAC':'Tyr','UAA':'STOP','UAG':'STOP',
       'CAU':'His','CAC':'His','CAA':'Gln','CAG':'Gln','AAU':'Asn','AAC':'Asn','AAA':'Lys','AAG':'Lys','GAU':'Asp',
       'GAC':'Asp','GAA':'Glu','GAG':'Glu','UGU':'Cys','UGC':'Cys','UGA':'STOP','UGG':'Trp','CGU':'Arg','CGC':'Arg',
       'CGA':'Arg','CGG':'Arg','AGU':'Ser','AGC':'Ser','AGA':'Arg','AGG':'Arg','GGU':'Gly','GGC':'Gly','GGA':'Gly',
       'GGG':'Gly'}

aa_counter = {
    'Pro' : [],
    'Ser' : [],
    'Ala' : [],
    'Thr' : [],
    'Leu' : [],
    'Phe' : [],
    'Val' : [],
    'Ile' : [],
    'Met' : [],
    'Arg' : [],
    'Cys' : [],
    'Trp' : [],
    'Gly' : [],
    'His' : [],
    'Gln' : [],
    'Tyr' : [],
    'Asp' : [],
    'Glu' : [],
    'Asn' : [],
    'Lys' : []
}

def create_genome(length=9999):
    generated_genome = ''.join([random.choice('AUGC') for n in range(length)])
    return generated_genome

#ref_data = pd.read_csv('../data/U_ideal_table.csv', index_col=0)

#ref_data = ref_data[ref_data['GenType'] == 'translated']

#ref_data = ref_data[['CodonNumber', 'RefCodon']].drop_duplicates()

#ref_data = ref_data[~(ref_data['RefCodon'].isin(['UGA', 'UAA', 'UAG']))]

#ref_codon_list = ref_data['RefCodon'].to_list()

#mutations = pd.read_csv('../data/All_mutation_information.csv.gz')
#mutations = mutations[mutations['(SBS) is coding'].isin([True, 'TRUE', 'TRUE, TRUE'])]

#all_df = mutations[mutations['base(s) in status 1'].isin(['A', 'T', 'G', 'C']) & mutations['base(s) in status 2'].isin(['A', 'T', 'G', 'C'])]

#all_df['aa_from'] = all_df['(SBS) AA change'].str.split(" ").str[1].str[0]
#all_df['aa_to'] = all_df['(SBS) AA change'].str.split(" ").str[1].str[0]
#all_df = all_df.reset_index(drop=True)
#all_df['counter'] = all_df.index
#all_df['from_to_nuc'] = all_df['base(s) in status 1'] + '>' + all_df['base(s) in status 2']
#data = all_df[['from_to_nuc', 'counter']].groupby(['from_to_nuc'], as_index=False).count()
#n_mut = int(sum(data['counter']))
#max_mut = int(max(data['counter']))
#data['counter'] = data['counter']/n_mut

#data['from_nuc'] = data['from_to_nuc'].str.split(">").str[0]
#data['to_nuc'] = data['from_to_nuc'].str.split(">").str[1]
#data['to_nuc'] = data['to_nuc'].str.replace('T','U')
#data['from_nuc'] = data['from_nuc'].str.replace('T','U')

#data

viruses_specs['from_to_nuc'] = viruses_specs['Mut']
viruses_specs['counter'] = viruses_specs['MutSpec']

viruses_specs['from_nuc'] = viruses_specs['from_to_nuc'].str.split(">").str[0]
viruses_specs['to_nuc'] = viruses_specs['from_to_nuc'].str.split(">").str[1]
viruses_specs['to_nuc'] = viruses_specs['to_nuc'].str.replace('T','U')
viruses_specs['from_nuc'] = viruses_specs['from_nuc'].str.replace('T','U')

viruses_specs = viruses_specs[['taxid', 'from_to_nuc', 'counter', 'from_nuc', 'to_nuc']]

viruses_specs.head()

def select_codon_number(codon_list):
    return random.choice(range(len(codon_list)))

def mutate_codon(codon_number, mutspec):
    nuc_number = random.choice(range(len(codon_list[codon_number])))
    
    nuc_spec = mutspec[mutspec['from_nuc'] == codon_list[codon_number][nuc_number]]
    pop = nuc_spec['to_nuc'].to_list()
    pop.append(codon_list[codon_number][nuc_number])
    
    wei = nuc_spec['counter'].to_list()
    wei.append(1 - sum(wei))
    
    new_nuc = random.choices(pop, weights=wei)[0]
    
    new_codon = codon_list[codon_number][:nuc_number] + new_nuc + codon_list[codon_number][nuc_number + 1:]
    
    if new_codon in ['UGA', 'UAA', 'UAG']:
        new_codon = codon_list[codon_number]
    
    codon_list[codon_number] = new_codon

def update_aa_dict(aa_counter, new_aa_counter):
    for key in aa_counter.keys():
        if key in new_aa_counter.keys():
            aa_counter[key].append(new_aa_counter[key])
        else:
            aa_counter[key].append(0)

viruses_g_l_dict = {
    'taxid' : [],
    'strand' : [],
    'gainers' : [],
    'loosers' : []
}

for tax in viruses_specs['taxid'].unique():
    aa_counter = {
    'Pro' : [],
    'Ser' : [],
    'Ala' : [],
    'Thr' : [],
    'Leu' : [],
    'Phe' : [],
    'Val' : [],
    'Ile' : [],
    'Met' : [],
    'Arg' : [],
    'Cys' : [],
    'Trp' : [],
    'Gly' : [],
    'His' : [],
    'Gln' : [],
    'Tyr' : [],
    'Asp' : [],
    'Glu' : [],
    'Asn' : [],
    'Lys' : []
    }
    
    virus_spec = viruses_specs[viruses_specs['taxid'] == tax]
    
    if tax in ('113201', '113202'):
        virus_aa_freq = viruses_aa_freq[viruses_aa_freq['taxid'] == '11320']
    else:
        virus_aa_freq = viruses_aa_freq[viruses_aa_freq['taxid'] == tax.split('_')[0]]
        
    print(tax)
    if tax in ('3052493', '1133363', '3052763'):
        strand = '-'
    else:
        virus_type = viruses_type[viruses_type['taxid'] == tax.replace('_','')]
        strand = virus_type['Type'].values[0]
        
    virus_name = virus_type['species'].values[0].split('__')[0]
    
    
    with PdfPages(f'../figures/simulations/mutation_of_aa_{tax}.pdf') as pdf:
        fig, axs = plt.subplots(5, 4, figsize=(20, 30))
        for _ in range(1):
            aa_counter = {
                'Pro' : [],
                'Ser' : [],
                'Ala' : [],
                'Thr' : [],
                'Leu' : [],
                'Phe' : [],
                'Val' : [],
                'Ile' : [],
                'Met' : [],
                'Arg' : [],
                'Cys' : [],
                'Trp' : [],
                'Gly' : [],
                'His' : [],
                'Gln' : [],
                'Tyr' : [],
                'Asp' : [],
                'Glu' : [],
                'Asn' : [],
                'Lys' : []
            }
            genome = create_genome(length=9999)
            codon_list = [genome[i:i+3] for i in range(0, len(genome), 3)]

            num_generations = 10000
            mut_in_gen = 100

            for gen in tqdm(range(num_generations)):
                new_dict = dict(Counter([codon_aa.get(item, item)  for item in codon_list]))

                update_aa_dict(aa_counter, new_dict)
                for mut in range(mut_in_gen):
                    codon_num_to_mut = select_codon_number(codon_list)

                    mutate_codon(codon_num_to_mut, mutspec=virus_spec)
            mutated_df = pd.DataFrame(aa_counter)
            mutated_df = mutated_df / len(codon_list)

            gainers = []
            loosers = []
            for aa in mutated_df.columns:
                last_freq = mutated_df[aa].tail(1).values[0]
                ref_freq = virus_aa_freq[aa].values[0]
                if last_freq>ref_freq:
                    gainers.append(aa)
                elif last_freq<ref_freq:
                    loosers.append(aa)
                    
            g_l = list(aa_counter.keys())
            column=0
            for aa_num in range(len(g_l)):
                aa = g_l[aa_num]
                ref_aa_count = virus_aa_freq[aa].values[0]
                if aa in gainers: 
                    colr = 'red'
                elif aa in loosers: 
                    colr = 'blue'

                if aa_num<=4:
                    axs[aa_num][column].plot(mutated_df[aa], color=colr, label=aa)
                    #axs[aa_num][column].legend(loc="upper right", fontsize=20)
                    axs[aa_num][column].yaxis.set_major_locator(plt.MaxNLocator(2))
                    axs[aa_num][column].yaxis.set_tick_params(labelsize=15)
                    axs[aa_num][column].axhline(y = ref_aa_count, color = 'black', linestyle = '--')
                elif aa_num>4 and aa_num<=9:
                    axs[aa_num-5][column+1].plot(mutated_df[aa], color=colr, label=aa)
                    #axs[aa_num-5][column+1].legend(loc="upper right", fontsize=20)
                    axs[aa_num-5][column+1].yaxis.set_major_locator(plt.MaxNLocator(2))
                    axs[aa_num-5][column+1].yaxis.set_tick_params(labelsize=15)
                    axs[aa_num-5][column+1].axhline(y = ref_aa_count, color = 'black', linestyle = '--')
                elif aa_num>9 and aa_num<=14:
                    axs[aa_num-10][column+2].plot(mutated_df[aa], color=colr, label=aa)
                    #axs[aa_num-10][column+2].legend(loc="upper right", fontsize=20)
                    axs[aa_num-10][column+2].yaxis.set_major_locator(plt.MaxNLocator(2))
                    axs[aa_num-10][column+2].yaxis.set_tick_params(labelsize=15)
                    axs[aa_num-10][column+2].axhline(y = ref_aa_count, color = 'black', linestyle = '--')
                elif aa_num>14:
                    axs[aa_num-15][column+3].plot(mutated_df[aa], color=colr, label=aa)
                    #axs[aa_num-15][column+3].legend(loc="upper right", fontsize=20)
                    axs[aa_num-15][column+3].yaxis.set_major_locator(plt.MaxNLocator(2))
                    axs[aa_num-15][column+3].yaxis.set_tick_params(labelsize=15)
                    axs[aa_num-15][column+3].axhline(y = ref_aa_count, color = 'black', linestyle = '--')

        axs[4][0].xaxis.set_tick_params(labelsize=15)
        axs[4][1].xaxis.set_tick_params(labelsize=15)
        axs[4][2].xaxis.set_tick_params(labelsize=15)
        axs[4][3].xaxis.set_tick_params(labelsize=15)
        plt.subplots_adjust(hspace=0)
        plt.subplots_adjust(wspace=0.15)
        
        for row in range(5):
            for column in range(4):
                handles, labels = axs[row][column].get_legend_handles_labels()
                by_label = dict(zip(labels, handles))
                axs[row][column].legend(by_label.values(), by_label.keys(), loc="upper right", fontsize=20)
                #axs[row][column].legend(loc="upper right", fontsize=20)

        plt.grid(False)
        fig.supylabel("Proportion of Amino Acid in the genome", x=0.08, size=20)
        fig.supxlabel("Number of Generations", y=0.09, size=20)
        fig.suptitle("Mutation of the amino acid composition of according to the mutation spectrum"+virus_name+" Genome type - "+strand, y=0.90, size=23)
        pdf.savefig(fig, bbox_inches='tight')
    plt.close()
    virus_g_l_dict = {
        'taxid' : tax,
        'strand' : strand,
        'gainers' : gainers,
        'loosers' : loosers
    }
    update_aa_dict(viruses_g_l_dict, virus_g_l_dict)
    
viruses_g_l_df = pd.DataFrame(viruses_g_l_dict)
viruses_g_l_df.to_csv('../data_obtain/viruses_g_l.csv')  

viruses_specs['taxid'].unique()

gainers = ['Ser', 'Leu', 'Phe', 'Ile', 'Cys', 'Tyr']
loosers = ['Pro', 'Ala', 'Thr', 'Val', 'Met', 'Arg', 'Trp', 'Gly', 'His', 'Gln', 'Asp', 'Glu', 'Asn', 'Lys']
with PdfPages('../figures/mutation_of_aa.pdf') as pdf:
    fig, axs = plt.subplots(7, 3, figsize=(20, 30))

    for aa_num in range(len(loosers)):
        if aa_num <= 6:
            first = aa_num
            second = 1
        else:
            first = aa_num - 7
            second = 2
        aa = loosers[aa_num]
        if aa in gainers:
            axs[first][second].plot(mutated_df[aa], color='red', label=aa)
        elif aa in loosers:
            axs[first][second].plot(mutated_df[aa], color='blue', label=aa)
        else:
            axs[first][second].plot(mutated_df[aa], color='black', label=aa)

        axs[first][second].legend(loc="upper right", fontsize=20)
        axs[first][second].yaxis.set_major_locator(plt.MaxNLocator(2))
        axs[first][second].yaxis.set_tick_params(labelsize=15)
        
    for aa_num in range(len(gainers)):
        aa = gainers[aa_num]
        if aa in gainers:
            axs[aa_num][0].plot(mutated_df[aa], color='red', label=aa)
        elif aa in loosers:
            axs[aa_num][0].plot(mutated_df[aa], color='blue', label=aa)
        else:
            axs[aa_num][0].plot(mutated_df[aa], color='black', label=aa)

        axs[aa_num][0].legend(loc="upper right", fontsize=20)
        axs[aa_num][0].yaxis.set_major_locator(plt.MaxNLocator(2))
        axs[aa_num][0].yaxis.set_tick_params(labelsize=15)
        
    axs[5][0].xaxis.set_tick_params(labelsize=15)
    axs[6][1].xaxis.set_tick_params(labelsize=15)
    axs[6][2].xaxis.set_tick_params(labelsize=15)
    plt.subplots_adjust(hspace=0)
    plt.subplots_adjust(wspace=0.15)
    axs[6,0].set_axis_off()
        
    plt.grid(False)
    fig.supylabel("Proportion of Amino Acid in the genome", x=0.08, size=20)
    fig.supxlabel("Number of Generations", y=0.09, size=20)
    fig.suptitle("Mutation of the amino acid composition of SARS-CoV-2 according to the mutation spectrum", y=0.90, size=23)
    pdf.savefig(fig, bbox_inches='tight')
